requirements.txt
mss~=6.0.0
numpy==1.16.6
Pillow~=7.0.0
psutil==5.8.0
PyInstaller==4.2
PyQt5==5.13.0
ujson~=4.0.1
qt-material==2.7
matplotlib~=3.2.2
opencv-python~=4.2.0.34